🐛 Fix loading of `UserFormTeamsCell.mock`
