+++ web/src/components/Admin/User/UserFormTeamsCell/UserFormTeamsCell.mock.ts
@@ -1,0 +1,22 @@
+export const standard = () => ({
+  teams: [
+    {
+      id: '1',
+      name: 'team1',
+    },
+    {
+      id: '2',
+      name: 'team2',
+    },
+  ],
+  roles: [
+    {
+      id: '3',
+      name: 'foo_role',
+    },
+    {
+      id: '4',
+      name: 'admin',
+    },
+  ],
+})
