--- web/src/components/Admin/User/UserFormTeamsCell/UserFormTeamsCell.mocks.ts
@@ -1,24 +1,0 @@
-export const standard = () => ({
-  userFormTeams: {
-    teams: [
-      {
-        id: '1',
-        name: 'team1',
-      },
-      {
-        id: '2',
-        name: 'team2',
-      },
-    ],
-    roles: [
-      {
-        id: '3',
-        name: 'foo_role',
-      },
-      {
-        id: '4',
-        name: 'admin',
-      },
-    ],
-  },
-})
